requirements.txt
# Core LLM and vector store
langchain>=0.1.15
langchain-community>=0.3.0
openai>=1.3.5
pinecone>=3.0.0

# Document processing
unstructured>=0.9.0
pdfplumber>=0.10.3     # optional: for better PDF extraction
PyMuPDF>=1.23.1         # optional: for PDF loading with fitz
tiktoken>=0.5.1
python-docx>=0.8.11     # for DOCX file processing

# Embeddings and ML
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0

# Web app (optional for UI)
streamlit>=1.31.1

# Utilities
python-dotenv>=1.0.0    # for managing environment variables

# If using async API calls or FastAPI server
fastapi>=0.110.0
uvicorn>=0.24.0

# Optional enhancements
pypdf>=3.17.1           # Alternate PDF loader
